documentation/requirements.txt
flask==3.0.0
httpx==0.26.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
mcp-client-python 